+++ demo/src/main/java/com/orange/ods/demo/ui/components/Component.kt
@@ -81,7 +81,7 @@ sealed class Component(
         R.drawable.il_lists,
         null,
         R.string.component_lists_description,
-        listOf(Variant.ListsOneLine, Variant.ListsTwoLines, Variant.ListsThreeLines)
+        listOf(Variant.Lists)
     )
 
     object Progress : Component(R.string.component_progress, R.drawable.il_progress, null, R.string.component_progress_description)
@@ -139,9 +139,7 @@ sealed class Variant(
     object Chip : Variant(R.string.component_chip)
     object ChipFilter : Variant(R.string.component_chip_filter)
 
-    object ListsOneLine : Variant(R.string.component_lists_one_line)
-    object ListsTwoLines : Variant(R.string.component_lists_two_lines)
-    object ListsThreeLines : Variant(R.string.component_lists_three_lines)
+    object Lists : Variant(R.string.component_lists_demo)
 
     object TextFieldsFilled : Variant(R.string.component_text_fields_filled)
     object TextFieldsOutlined : Variant(R.string.component_text_fields_outlined)