+++ demo/src/main/java/com/orange/ods/demo/ui/components/VariantDetailScreen.kt
@@ -18,7 +18,7 @@ import com.orange.ods.demo.ui.components.buttons.ComponentButtons
 import com.orange.ods.demo.ui.components.cards.ComponentCard
 import com.orange.ods.demo.ui.components.chips.VariantChip
 import com.orange.ods.demo.ui.components.chips.VariantChipFilter
-import com.orange.ods.demo.ui.components.lists.ComponentList
+import com.orange.ods.demo.ui.components.lists.ComponentLists
 import com.orange.ods.demo.ui.components.tabs.ComponentTabs
 import com.orange.ods.demo.ui.components.tabs.TabsConfiguration
 import com.orange.ods.demo.ui.components.textfields.ComponentTextField
@@ -40,7 +40,7 @@ fun VariantDetailScreen(
             Component.Buttons -> ComponentButtons(variant = variant)
             Component.Cards -> ComponentCard(variant = variant)
             Component.Chips -> if (variant == Variant.ChipFilter) VariantChipFilter() else VariantChip()
-            Component.Lists -> ComponentList(variant = variant)
+            Component.Lists -> ComponentLists()
             Component.TextFields -> ComponentTextField(variant = variant)
             Component.Tabs -> ComponentTabs(variant, updateTopAppBarTabs)
             else -> {}