+++ lib/src/main/java/com/orange/ods/compose/component/list/OdsListItem.kt
@@ -20,6 +20,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.requiredHeight
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.material.Divider
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Icon
 import androidx.compose.material.ListItem
@@ -35,7 +36,11 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Dp
 import com.orange.ods.R
+import com.orange.ods.compose.component.utilities.OdsImageCircleShape
 import com.orange.ods.compose.text.OdsTextSubtitle1
+import com.orange.ods.utilities.extension.getElementOfType
+import com.orange.ods.utilities.extension.isNotNullOrBlank
+import com.orange.ods.utilities.extension.orElse
 
 /**
  * <a href="https://system.design.orange.com/0c1af118d/p/09a804-lists/b/669743" target="_blank">ODS Lists</a>.
@@ -43,6 +48,8 @@ import com.orange.ods.compose.text.OdsTextSubtitle1
  * Lists are continuous, vertical indexes of text or images.
  *
  * To make this [OdsListItem] clickable, use [Modifier.clickable].
+ *
+ * To specify an icon type, use [Modifier.iconType] on [modifier] and call [OdsListItemIconScope.OdsListItemIcon] in the [icon] lambda.
  *
  * This component can be used to achieve the list item templates existing in the spec. For example:
  * - one-line items
@@ -52,12 +59,9 @@ import com.orange.ods.compose.text.OdsTextSubtitle1
  * - three-line items
  * @sample androidx.compose.material.samples.ThreeLineListItems
  *
- * Note: If you want to display a big thumbnail without left margin in your items, please use [OdsListItemWideThumbnail].
- *
  * @param modifier Modifier to be applied to the list item
  * @param text The primary text of the list item
  * @param icon The leading supporting visual of the list item
- * @param isThumbnailIcon Whether the icon is a thumbnail (more space must be allowed in this case)
  * @param secondaryText The secondary text of the list item
  * @param singleLineSecondaryText Whether the secondary text is single line
  * @param overlineText The text displayed above the primary text
@@ -68,18 +72,70 @@ import com.orange.ods.compose.text.OdsTextSubtitle1
 fun OdsListItem(
     modifier: Modifier = Modifier,
     text: String,
-    icon: @Composable (() -> Unit)? = null,
-    isThumbnailIcon: Boolean = false,
+    icon: @Composable (OdsListItemIconScope.() -> Unit)? = null,
     secondaryText: String? = null,
     singleLineSecondaryText: Boolean = true,
     overlineText: String? = null,
     trailing: @Composable (() -> Unit)? = null
 ) {
+    val iconType = modifier.getElementOfType<OdsListItemIconTypeModifier>()?.iconType
+    val listItemIconScope = OdsListItemIconScope(iconType)
+    if (iconType == OdsListItemIconType.WideImage) {
+        Row(modifier = modifier, verticalAlignment = Alignment.CenterVertically) {
+            icon?.let { listItemIconScope.it() }
+            OdsListItemInternal(
+                modifier = Modifier
+                    .weight(1f)
+                    .iconType(OdsListItemIconType.WideImage),
+                listItemScope = listItemIconScope,
+                text = text,
+                icon = null,
+                secondaryText = secondaryText,
+                singleLineSecondaryText = singleLineSecondaryText,
+                overlineText = overlineText,
+                trailing = trailing
+            )
+        }
+    } else {
+        OdsListItemInternal(
+            modifier = modifier,
+            listItemScope = listItemIconScope,
+            text = text,
+            icon = icon,
+            secondaryText = secondaryText,
+            singleLineSecondaryText = singleLineSecondaryText,
+            overlineText = overlineText,
+            trailing = trailing
+        )
+    }
+
+    modifier.getElementOfType<OdsListItemDividerModifier>()?.let { dividerModifier ->
+        val startIndent = dividerModifier.startIndent
+            .orElse { iconType?.getDividerStartIndent() }
+            .orElse { dimensionResource(id = R.dimen.spacing_m) }
+        Divider(startIndent = startIndent)
+    }
+}
+
+@ExperimentalMaterialApi
+@Composable
+private fun OdsListItemInternal(
+    modifier: Modifier = Modifier,
+    listItemScope: OdsListItemIconScope,
+    text: String,
+    icon: @Composable (OdsListItemIconScope.() -> Unit)? = null,
+    secondaryText: String? = null,
+    singleLineSecondaryText: Boolean = true,
+    overlineText: String? = null,
+    trailing: @Composable (() -> Unit)? = null
+) {
+    val iconType = modifier.getElementOfType<OdsListItemIconTypeModifier>()?.iconType
     val requiredHeight = computeRequiredHeight(
         hasIcon = icon != null,
-        isThumbnailIcon = isThumbnailIcon,
-        hasOverline = overlineText != null,
-        hasSecondaryText = secondaryText != null,
+        iconType = iconType,
+        hasOverlineText = overlineText.isNotNullOrBlank(),
+        hasText = text.isNotBlank(),
+        hasSecondaryText = secondaryText.isNotNullOrBlank(),
         singleLineSecondaryText = singleLineSecondaryText
     )
     val secondaryTextLinesNumber = if (singleLineSecondaryText || (overlineText != null && secondaryText != null)) 1 else 2
@@ -87,108 +143,64 @@ fun OdsListItem(
         modifier = modifier
             .fillMaxWidth()
             .requiredHeight(requiredHeight),
-        icon = icon,
-        secondaryText = if (secondaryText != null) {
+        icon = icon?.let { { listItemScope.it() } },
+        secondaryText = if (secondaryText.isNotNullOrBlank()) {
             { Text(text = secondaryText, style = MaterialTheme.typography.body2, maxLines = secondaryTextLinesNumber, overflow = TextOverflow.Ellipsis) }
         } else null,
         singleLineSecondaryText = singleLineSecondaryText,
-        overlineText = if (overlineText != null) {
+        overlineText = if (overlineText.isNotNullOrBlank()) {
             { Text(text = overlineText, style = MaterialTheme.typography.overline, color = MaterialTheme.colors.onSurface.copy(alpha = 0.6f)) }
         } else null,
         trailing = trailing,
         text = {
-            OdsTextSubtitle1(text = text)
+            if (text.isNotBlank()) {
+                OdsTextSubtitle1(text = text)
+            }
         }
     )
 }
 
 /**
- * Use this list item to display a wide thumbnail without start padding in the item, otherwise use [OdsListItem].
+ * Displays an icon in a list item.
  *
- * To make this [OdsListItemWideThumbnail] clickable, use [Modifier.clickable].
+ * This method throws an exception if no icon type has been specified on the [OdsListItem] modifier using the [Modifier.iconType] method.
  *
- * @param modifier Modifier to be applied to the list item
- * @param text The primary text of the list item
- * @param thumbnail The painter of the thumbnail
- * @param thumbnailContentDescription The content description for the given thumbnail
- * @param secondaryText The secondary text of the list item
- * @param singleLineSecondaryText Whether the secondary text is single line
- * @param overlineText The text displayed above the primary text
- * @param trailing The trailing meta text, icon, switch or checkbox
+ * @param painter to draw
+ * @param contentDescription Content description of the icon
  */
-@ExperimentalMaterialApi
 @Composable
-fun OdsListItemWideThumbnail(
-    modifier: Modifier = Modifier,
-    text: String,
-    thumbnail: Painter,
-    thumbnailContentDescription: String? = null,
-    secondaryText: String? = null,
-    singleLineSecondaryText: Boolean = true,
-    overlineText: String? = null,
-    trailing: @Composable (() -> Unit)? = null
-) {
-    Row(modifier = modifier, verticalAlignment = Alignment.CenterVertically) {
-        OdsListWideThumbnail(painter = thumbnail, contentDescription = thumbnailContentDescription)
-        OdsListItem(
-            modifier = Modifier.weight(1f),
-            isThumbnailIcon = true,
-            text = text,
-            secondaryText = secondaryText,
-            singleLineSecondaryText = singleLineSecondaryText,
-            overlineText = overlineText,
-            trailing = trailing
-        )
+fun OdsListItemIconScope.OdsListItemIcon(painter: Painter, contentDescription: String? = null) {
+    when (iconType) {
+        OdsListItemIconType.Icon -> {
+            Column(modifier = Modifier.fillMaxHeight(), verticalArrangement = Arrangement.Center) {
+                Icon(painter = painter, contentDescription = contentDescription)
+            }
+        }
+        OdsListItemIconType.CircularImage -> {
+            OdsImageCircleShape(painter = painter, contentDescription = contentDescription)
+        }
+        OdsListItemIconType.SquareImage -> {
+            Image(
+                painter = painter,
+                contentDescription = contentDescription,
+                modifier = Modifier
+                    .size(dimensionResource(id = R.dimen.list_square_image_size))
+                    .clip(MaterialTheme.shapes.medium),
+                contentScale = ContentScale.Crop
+            )
+        }
+        OdsListItemIconType.WideImage -> {
+            Image(
+                painter = painter,
+                contentDescription = contentDescription,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier
+                    .height(dimensionResource(id = R.dimen.list_wide_image_height))
+                    .width(dimensionResource(id = R.dimen.list_wide_image_width))
+            )
+        }
+        null -> throw Exception("OdsListItemIcon(Painter, String?) method has been called without specifying an icon type. Please specify an icon type by calling the Modifier.iconType(OdsListItemIcon) method on the OdsList modifier.")
     }
-}
-
-/**
- * Displays an icon in a list item centered vertically
- *
- * @param painter to draw
- * @param contentDescription Content description of the icon
- */
-@Composable
-fun OdsListItemIcon(painter: Painter, contentDescription: String? = null) {
-    Column(modifier = Modifier.fillMaxHeight(), verticalArrangement = Arrangement.Center) {
-        Icon(painter = painter, contentDescription = contentDescription)
-    }
-}
-
-/**
- * Displays a 56x100 thumbnail in a list item
- *
- * @param painter to draw
- * @param contentDescription Content description of the icon
- */
-@Composable
-fun OdsListWideThumbnail(painter: Painter, contentDescription: String? = null) {
-    Image(
-        painter = painter,
-        contentDescription = contentDescription,
-        contentScale = ContentScale.Crop,
-        modifier = Modifier
-            .height(dimensionResource(id = R.dimen.list_wide_thumbnail_height))
-            .width(dimensionResource(id = R.dimen.list_wide_thumbnail_width))
-    )
-}
-
-/**
- * Displays a 56x56 thumbnail in a list item
- *
- * @param painter to draw
- * @param contentDescription Content description of the icon
- */
-@Composable
-fun OdsListSquaredThumbnail(painter: Painter, contentDescription: String? = null) {
-    Image(
-        painter = painter,
-        contentDescription = contentDescription,
-        modifier = Modifier
-            .size(dimensionResource(id = R.dimen.list_squared_thumbnail_size))
-            .clip(MaterialTheme.shapes.medium),
-        contentScale = ContentScale.Crop
-    )
 }
 
 /**
@@ -196,21 +208,104 @@ fun OdsListSquaredThumbnail(painter: Painter, contentDescription: String? = null
  * It allows to be able to center vertically elements in the item.
  */
 @Composable
-internal fun computeRequiredHeight(hasIcon: Boolean, isThumbnailIcon: Boolean, hasOverline: Boolean, hasSecondaryText: Boolean, singleLineSecondaryText: Boolean): Dp {
+private fun computeRequiredHeight(
+    hasIcon: Boolean,
+    iconType: OdsListItemIconType?,
+    hasOverlineText: Boolean,
+    hasText: Boolean,
+    hasSecondaryText: Boolean,
+    singleLineSecondaryText: Boolean
+): Dp {
+    val wideImage = iconType == OdsListItemIconType.WideImage
     val heightRes = when {
         // single-line
-        !hasOverline && !hasSecondaryText -> when {
-            hasIcon && !isThumbnailIcon -> R.dimen.list_single_line_with_icon_item_height
-            isThumbnailIcon -> R.dimen.list_single_line_with_thumbnail_item_height
+        !hasOverlineText && (!hasSecondaryText || !hasText) -> when {
+            hasIcon && !wideImage -> R.dimen.list_single_line_with_icon_item_height
+            wideImage -> R.dimen.list_single_line_with_wide_image_item_height
             else -> R.dimen.list_single_line_item_height
         }
         // three-lines
-        hasOverline && hasSecondaryText -> R.dimen.list_three_line_item_height
+        hasOverlineText && hasSecondaryText -> R.dimen.list_three_line_item_height
         // two-lines
-        hasOverline || (hasSecondaryText && singleLineSecondaryText) -> if (hasIcon || isThumbnailIcon) R.dimen.list_two_line_item_with_icon_height else R.dimen.list_two_line_item_height
+        hasOverlineText || (hasSecondaryText && singleLineSecondaryText) -> if (hasIcon || wideImage) R.dimen.list_two_line_with_icon_item_height else R.dimen.list_two_line_item_height
         // three-lines
         else -> R.dimen.list_three_line_item_height
     }
 
     return dimensionResource(id = heightRes)
+}
+
+/**
+ * An [OdsListItemIconScope] provides a scope for the icon of [OdsListItem].
+ *
+ * @param iconType The icon type
+ */
+data class OdsListItemIconScope(val iconType: OdsListItemIconType?)
+
+/**
+ * Represents the various types of icon that can be displayed in an [OdsListItem].
+ */
+enum class OdsListItemIconType {
+
+    /** A standard icon. */
+    Icon,
+
+    /** An image cropped into a circle. */
+    CircularImage,
+
+    /** An image cropped into a square. */
+    SquareImage,
+
+    /** An image cropped into a rectangle. */
+    WideImage
+}
+
+@Composable
+private fun OdsListItemIconType.getDividerStartIndent(): Dp {
+    return when (this) {
+        OdsListItemIconType.Icon,
+        OdsListItemIconType.CircularImage -> dimensionResource(id = R.dimen.avatar_size) + dimensionResource(id = R.dimen.spacing_m).times(2)
+        OdsListItemIconType.SquareImage -> dimensionResource(id = R.dimen.list_square_image_size) + dimensionResource(id = R.dimen.spacing_m).times(2)
+        OdsListItemIconType.WideImage -> dimensionResource(id = R.dimen.list_wide_image_width) + dimensionResource(id = R.dimen.spacing_m)
+    }
+}
+
+/**
+ * Specifies the icon type to display in an [OdsListItem].
+ *
+ * @param iconType The icon type
+ */
+fun Modifier.iconType(iconType: OdsListItemIconType): Modifier {
+    return then(object : OdsListItemIconTypeModifier {
+        override val iconType: OdsListItemIconType
+            get() = iconType
+    })
+}
+
+/**
+ * Displays a divider at the bottom of an [OdsListItem].
+ *
+ * @param startIndent The start indent of the divider
+ */
+fun Modifier.divider(startIndent: Dp? = null): Modifier {
+    return then(object : OdsListItemDividerModifier {
+        override val startIndent: Dp?
+            get() = startIndent
+    })
+}
+
+/**
+ * A modifier that allows to configure the icon type in an [OdsListItem].
+ */
+private interface OdsListItemIconTypeModifier : Modifier.Element {
+
+    val iconType: OdsListItemIconType
+}
+
+/**
+ * A modifier that allows to display a divider at the bottom of an [OdsListItem].
+ */
+private interface OdsListItemDividerModifier : Modifier.Element {
+
+    val startIndent: Dp?
 }