+++ lib/src/main/java/com/orange/ods/utilities/extension/ModifierExt.kt
@@ -25,3 +25,15 @@ inline fun Modifier.noRippleClickable(crossinline onClick: () -> Unit): Modifier
         onClick()
     }
 }
+
+/**
+ * Returns the first [Modifier.Element] of type [T] in the current modifier, or null if an element of type [T] could not be found.
+ *
+ * @param T The type of the [Modifier.Element].
+ * @return The modifier element, or null if it could not be found.
+ */
+internal inline fun <reified T> Modifier.getElementOfType(): T? where T : Modifier.Element {
+    return foldOut(null as T?) { currentElement, foundElement ->
+        foundElement.orElse { currentElement as? T }
+    }
+}